rewrite frontend and stage 1 compilation
with new design, the concrete syntax of display
is more natural (see test & test2.xml for example)
where a tree hierarchy of xml elements is used
to represent the model.

As we use a tree hierarchy (with list), the technique
to overcome coq's decreasing argument rule has
been learned and applied.

The first stage of the compiler is to generate global variable
definitions and infer type from the Lustre backend

The makefile has also been modified according to compcert ver.3
with better automation